Transform garden cluster bootstrapping into runnable
